Changed getFighter___No Returns to u32
No reason actually to return these as u8, their numeric ranges are already enforced by where we're getting them from, we don't need to worry about that at this point.

+++ lavaGameTweaks/include/fighterHooks.h
@@ -40,8 +40,8 @@ namespace fighterHooks
 
 #define INCLUDE_OUTSIDE_OBSERVER false
 
-	u8 getFighterSlotNo(Fighter* fighterIn);
-	u8 getFighterPlayerNo(Fighter* fighterIn);
+	u32 getFighterSlotNo(Fighter* fighterIn);
+	u32 getFighterPlayerNo(Fighter* fighterIn);
 
 	class ftCallbackMgr
 	{
